[configuration] split loaders into own modules
so they can evolve independently and start form their api
